Fix #6542: some checks for sprite numbers >= 65536 failed due to the sprite number being stored in a short

+++ src/act123_classes.cpp
@@ -62,7 +62,7 @@ uint act123::MaxFoundFeat()const{
 
 void act123::IDarray::define(uint feature,unsigned int id,bool checks1C){
 	_m[id].used=false;
-	_m[id].sprite=(unsigned short)_spritenum;
+	_m[id].sprite=_spritenum;
 	_m[id].v1C=checks1C;
 	_m[id].feature=(ushort)feature;
 }
